+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.2/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] - 2025-10-10
@@ -16,3 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.2] - 2025-10-10
 ### Changed
 - more JS examples in README.md 
+
+## [1.0.3] - 2025-11-02
+### Fixed
+- Support comma-separated selectors with relative combinators (`>`, `~`, `+`)  
+
+### Changed
+- link to InDom GitHub Discussions in README.md Contribute section